requirements.txt
streamlit
tensorflow-cpu
xgboost
numpy
scikit-learn
pandas